Support weak references
The patch adds support for weak references. The original MinGW
implementation targets ix86, which handles weak symbols differently
compared to AArch64. In AArch64, the weak symbols are replaced by
other symbols which reference the original weak symbols, and the
compiler does not track the original symbol names.
This patch resolves this and declares the original symbols.

Here is an explanation of why this change is needed and what the
difference is between x86_64-w64-mingw32 and aarch64-w64-mingw32.

The way x86_64 calls a weak function:
call  weak_fn2

GCC emits the call and creates the required definitions at the end
of the assembly:

.weak weak_fn2
.def  weak_fn2;   .scl  2;    .type 32;   .endef

This is different from aarch64:

weak_fn2 will be legitimized and replaced by .refptr.weak_fn2,
and there will be no other references to weak_fn2 in the code.

adrp  x0, .refptr.weak_fn2
add   x0, x0, :lo12:.refptr.weak_fn2
ldr   x0, [x0]
blr   x0

GCC does not emit the required definitions at the end of the assembly,
and weak_fn2 is tracked only by the mingw stub sybmol.

Without the change, the stub definition will emit:

    .section      .rdata$.refptr.weak_fn2, "dr"
    .globl  .refptr.weak_fn2
    .linkonce     discard
.refptr.weak_fn2:
    .quad   weak_fn2

which is not enough. This fix will emit the required definitions:

    .weak   weak_fn2
    .def    weak_fn2;   .scl  2;    .type 32;   .endef
    .section      .rdata$.refptr.weak_fn2, "dr"
    .globl  .refptr.weak_fn2
    .linkonce     discard
.refptr.weak_fn2:
    .quad   weak_fn2

gcc/ChangeLog:

	* config/aarch64/cygming.h (SUB_TARGET_RECORD_STUB): Request
	declaration for weak symbols.
	(PE_COFF_LEGITIMIZE_EXTERN_DECL): Legitimize external
	declaration for weak symbols.
	* config/i386/cygming.h (SUB_TARGET_RECORD_STUB): Update
	declarations in ix86 with the same functionality.
	(PE_COFF_LEGITIMIZE_EXTERN_DECL): Likewise.
	* config/mingw/winnt-dll.cc (legitimize_pe_coff_symbol):
	Support declaration for weak symbols if requested.
	* config/mingw/winnt.cc (struct stub_list): Likewise.
	(mingw_pe_record_stub): Likewise.
	(mingw_pe_file_end): Likewise.
	* config/mingw/winnt.h (mingw_pe_record_stub): Likewise.